Rename parser.ts to amfParser.ts
