+++ src/common/amfParser.ts
