+++ src/index.ts
@@ -5,7 +5,7 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 export * as amf from "amf-client-js";
-export * as amfParser from "./common/parser";
+export * as amfParser from "./common/amfParser";
 export * as diff from "./diff";
 export * as download from "./download";
 export * as generate from "./generate";
